[BUGFIX] Prevent SiteNotFoundException in page module
When accessing a page with the page module that is not inside
a configured site you currently get an exception:

No site found in root line of page xxx.

This Exception is now catched and the inclusion of the
Siteimporve JS in the page module is skipped when no
domain is available.